add fix_gox.go
