+++ gopls/internal/goxls/imports/fix.go
@@ -1484,6 +1484,7 @@ func loadExportsFromFiles(ctx context.Context, env *ProcessEnv, dir string, incl
 	var pkgName string
 	var exports []string
 	fset := token.NewFileSet()
+	// goxls: scope list for check GopPackage
 	var scopes []*ast.Scope
 	for _, fi := range files {
 		select {
@@ -1509,50 +1510,18 @@ func loadExportsFromFiles(ctx context.Context, env *ProcessEnv, dir string, incl
 			// x_test package. We want internal test files only.
 			continue
 		}
+		// goxls: save scope
 		scopes = append(scopes, f.Scope)
 		pkgName = f.Name.Name
 	}
 	// goxls: export Go+ style func, startLower and overload (GopPackage)
-	var gopPackage bool
-	for _, scope := range scopes {
-		if obj := scope.Lookup("GopPackage"); obj != nil && obj.Kind == ast.Con {
-			gopPackage = true
-			break
-		}
-	}
-	for _, scope := range scopes {
-		for name, obj := range scope.Objects {
-			if ast.IsExported(name) {
-				exports = append(exports, name)
-				switch obj.Kind {
-				case ast.Fun:
-					if v, ok := toStartWithLowerCase(name); ok {
-						exports = append(exports, v)
-					}
-					if gopPackage && strings.HasSuffix(name, "__0") {
-						name = name[:len(name)-3]
-						exports = append(exports, name)
-						if v, ok := toStartWithLowerCase(name); ok {
-							exports = append(exports, v)
-						}
-					}
-				}
-			}
-		}
-	}
+	exports = gopExports(scopes)
 	if env.Logf != nil {
 		sortedExports := append([]string(nil), exports...)
 		sort.Strings(sortedExports)
 		env.Logf("loaded exports in dir %v (package %v): %v", dir, pkgName, strings.Join(sortedExports, ", "))
 	}
 	return pkgName, exports, nil
-}
-
-func toStartWithLowerCase(name string) (string, bool) {
-	if c := name[0]; c >= 'A' && c <= 'Z' {
-		return string(c+('a'-'A')) + name[1:], true
-	}
-	return name, false
 }
 
 // findImport searches for a package with the given symbols.
